Cleaned up intcode input, converted intcode to use long instead of int

+++ 2019/lib/include/intcode.h
@@ -9,7 +9,7 @@ typedef struct IntcodeReturn
 {
     int pc;
     int input_used;
-    int retval;
+    long retval;
     int halt;
 } IntcodeReturn_t;
 
@@ -19,11 +19,11 @@ void print_program(int program_length, int *buffer);
 // Returns length of intcode program if success
 // Returns -1 if file is invalid
 // Returns -2 if buffer is not long enough
-int read_intcode(const int bufsize, int *buffer, const char *filename);
+int read_intcode(const int bufsize, long *buffer, const char *filename);
 
 // Returns 0 if program terminated successfully (opcode 99)
 // Returns -1 if program terminated unexpectedly
-IntcodeReturn_t run_intcode(const int program_length, const int bufsize, int *buffer, const int input_length, int *input_buffer, int pc);
+IntcodeReturn_t run_intcode(const int program_length, const int bufsize, long *buffer, const int input_length, long *input_buffer, int pc);
 
 #ifdef __cplusplus
 }
