+++ 2019/sol/src/7.c
@@ -7,18 +7,18 @@
 #define NUM_AMPS 5
 #define LOOP_OFFSET 5
 
-static int run_amplifiers(int program_length, const int bufsize, int *program,
-                          int phase_settings[NUM_AMPS])
+static long run_amplifiers(int program_length, const int bufsize, long *program,
+                           long phase_settings[NUM_AMPS])
 {
     int previous_output = 0;
 
     for(int amp = 0; amp < NUM_AMPS; ++amp)
     {
-        int amp_program[bufsize];
-        memcpy(amp_program, program, bufsize * sizeof(int));
+        long amp_program[bufsize];
+        memcpy(amp_program, program, bufsize * sizeof(long));
 
         int num_inputs = 2;
-        int input_buffer[2] = {phase_settings[amp], previous_output};
+        long input_buffer[2] = {phase_settings[amp], previous_output};
 
         // intcode program isn't running through to halt, but it doesn't matter
         // since we're getting the output we need
@@ -30,8 +30,8 @@ static int run_amplifiers(int program_length, const int bufsize, int *program,
     return previous_output;
 }
 
-static int try_phase_settings(int depth, int phase_settings[NUM_AMPS],
-                              int program_length, const int bufsize, int *program)
+static long try_phase_settings(int depth, long phase_settings[NUM_AMPS],
+                               int program_length, const int bufsize, long *program)
 {
     if(depth == NUM_AMPS)
     {
@@ -55,8 +55,8 @@ static int try_phase_settings(int depth, int phase_settings[NUM_AMPS],
         if(!already_used)
         {
             phase_settings[depth] = i;
-            int max_with_phase = try_phase_settings(depth + 1, phase_settings,
-                                                    program_length, bufsize, program);
+            long max_with_phase = try_phase_settings(depth + 1, phase_settings,
+                                                     program_length, bufsize, program);
             if(max_with_phase > max_signal)
             {
                 max_signal = max_with_phase;
@@ -67,20 +67,20 @@ static int try_phase_settings(int depth, int phase_settings[NUM_AMPS],
     return max_signal;
 }
 
-int solve_2019_7_1(char const *input_filename)
+long solve_2019_7_1(char const *input_filename)
 {
     const int bufsize = 1000;
-    int buffer[bufsize];
+    long buffer[bufsize];
 
     int program_length = read_intcode(bufsize, buffer, input_filename);
     assert(program_length != -1);
 
-    int phase_settings[NUM_AMPS] = {0};
+    long phase_settings[NUM_AMPS] = {0};
     return try_phase_settings(0, phase_settings, program_length, bufsize, buffer);
 }
 
-static int run_amplifiers_loop(int program_length, const int bufsize, int *program,
-                               int phase_settings[NUM_AMPS])
+static long run_amplifiers_loop(int program_length, const int bufsize, long *program,
+                                long phase_settings[NUM_AMPS])
 {
     int previous_loop_end = 0;
     IntcodeReturn_t retval = {
@@ -90,11 +90,11 @@ static int run_amplifiers_loop(int program_length, const int bufsize, int *progr
         .halt = 0,
     };
 
-    int amp_programs[NUM_AMPS][bufsize];
-    int amp_pc[NUM_AMPS] = {0};
+    long amp_programs[NUM_AMPS][bufsize];
+    long amp_pc[NUM_AMPS] = {0};
     for(int amp = 0; amp < NUM_AMPS; ++amp)
     {
-        memcpy(amp_programs[amp], program, bufsize * sizeof(int));
+        memcpy(amp_programs[amp], program, bufsize * sizeof(long));
     }
 
     int first_iteration = 1;
@@ -104,7 +104,7 @@ static int run_amplifiers_loop(int program_length, const int bufsize, int *progr
         for(int amp = 0; amp < NUM_AMPS; ++amp)
         {
             int num_inputs = 1;
-            int input_buffer[2] = {previous_output, previous_output};
+            long input_buffer[2] = {previous_output, previous_output};
 
             if(first_iteration)
             {
@@ -130,15 +130,15 @@ static int run_amplifiers_loop(int program_length, const int bufsize, int *progr
     return previous_loop_end;
 }
 
-static int try_phase_settings_loop(int depth, int phase_settings[NUM_AMPS],
-                                   int program_length, const int bufsize, int *program)
+static long try_phase_settings_loop(int depth, long phase_settings[NUM_AMPS],
+                                    int program_length, const int bufsize, long *program)
 {
     if(depth == NUM_AMPS)
     {
         return run_amplifiers_loop(program_length, bufsize, program, phase_settings);
     }
 
-    int max_signal = 0;
+    long max_signal = 0;
 
     for(int i = LOOP_OFFSET; i < LOOP_OFFSET + NUM_AMPS; ++i)
     {
@@ -155,7 +155,7 @@ static int try_phase_settings_loop(int depth, int phase_settings[NUM_AMPS],
         if(!already_used)
         {
             phase_settings[depth] = i;
-            int max_with_phase = try_phase_settings_loop(depth + 1, phase_settings,
+            long max_with_phase = try_phase_settings_loop(depth + 1, phase_settings,
                                                          program_length, bufsize, program);
             if(max_with_phase > max_signal)
             {
@@ -167,14 +167,14 @@ static int try_phase_settings_loop(int depth, int phase_settings[NUM_AMPS],
     return max_signal;
 }
 
-int solve_2019_7_2(char const *input_filename)
+long solve_2019_7_2(char const *input_filename)
 {
     const int bufsize = 1000;
-    int buffer[bufsize];
+    long buffer[bufsize];
 
     int program_length = read_intcode(bufsize, buffer, input_filename);
     assert(program_length != -1);
 
-    int phase_settings[NUM_AMPS] = {0};
+    long phase_settings[NUM_AMPS] = {0};
     return try_phase_settings_loop(0, phase_settings, program_length, bufsize, buffer);
 }
